# backend/OpenApiSpec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Secure Message API",
    "version": "1.1.0",
    "description": "Backend API for the Secure Message System. The server stores only opaque encrypted message payloads."
  },
  "servers": [
    { "url": "http://localhost:8080" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "example": { "status": "ok" } } }
          }
        }
      }
    },
    "/api/register": {
      "post": {
        "summary": "Register user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password", "public_key", "encrypted_private_key"],
                "properties": {
                  "username": { "type": "string" },
                  "password": {
                    "type": "string",
                    "description": "8-128 characters with at least one number and one special character."
                  },
                  "public_key": { "type": "string" },
                  "encrypted_private_key": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Registered" },
          "400": { "description": "Invalid request" },
          "409": { "description": "User exists" }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token issued" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/logout": {
      "post": {
        "summary": "Logout user",
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": { "description": "Logged out" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/me": {
      "get": {
        "summary": "Current user info",
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": { "description": "User details" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/contacts": {
      "get": {
        "summary": "List saved contacts for the current user",
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": { "description": "Contacts list" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Save or update a contact",
        "security": [ { "bearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username"],
                "properties": {
                  "username": { "type": "string" },
                  "alias": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Contact saved" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/contacts/{alias}": {
      "delete": {
        "summary": "Remove a saved contact by alias",
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": { "description": "Contact removed" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Alias not found" }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "summary": "List all usernames (admin only)",
        "description": "Returns the registered usernames only. This admin view is separate from the messaging flow.",
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": { "description": "List of usernames" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/users/{username}/public-key": {
      "get": {
        "summary": "Get public key by username",
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Public key" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/messages": {
      "post": {
        "summary": "Send message",
        "description": "Stores an opaque encrypted payload. The ciphertext may decrypt to plain text or to a client-defined attachment envelope for arbitrary files, including images.",
        "security": [ { "bearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["recipient", "encrypted_key", "ciphertext", "iv"],
                "properties": {
                  "recipient": { "type": "string" },
                  "encrypted_key": { "type": "string" },
                  "ciphertext": { "type": "string" },
                  "iv": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Stored" },
          "401": { "description": "Unauthorized" }
        }
      },
      "get": {
        "summary": "List messages",
        "description": "Returns opaque encrypted message payloads. Clients are responsible for decrypting text or attachment envelopes locally.",
        "security": [ { "bearerAuth": [] } ],
        "parameters": [
          { "name": "with", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "before_id", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Messages list" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}